Remove BuySeekButton/SellOfferButton components, create button components and move them to sub folder, Adjust AddForm components
