--- src/js/components/BuySeekButton.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react';
-
-export default class BuySeekButton extends Component {
-  render(){
-    return(
-      <a href="/advertisements" className="uk-button uk-button-primary uk-button-large uk-width-1">Buy / Seek</a>
-    );
-  }
-}
