--- src/js/components/SellOfferButton.js
@@ -1,9 +1,0 @@
-import React, { Component } from 'react';
-
-export default class SellTradeButton extends Component {
-  render(){
-    return(
-      <a href="/advertisements" className="uk-button uk-button-danger uk-button-large uk-width-1">Sell / Offer</a>
-    );
-  }
-}
