+++ src/js/components/buttons/CancelButton.js
@@ -1,0 +1,10 @@
+import React, { Component } from 'react';
+import { browserHistory } from 'react-router';
+
+export default class CancelButton extends Component {
+  render(){
+    return(
+      <a onClick={browserHistory.goBack} className="uk-button uk-button-danger uk-button-large uk-width-1">Cancel</a>
+    );
+  }
+}
