+++ src/js/components/buttons/SubmitButton.js
@@ -1,0 +1,9 @@
+import React, { Component } from 'react';
+
+export default class SubmitButton extends Component {
+  render(){
+    return(
+      <button className="button-success uk-button uk-button-large uk-width-1" type="submit" value="Submit">Submit</button>
+    );
+  }
+}
