+++ src/js/pages/UserAdvertisementListPage.js
@@ -3,7 +3,7 @@ import api from '../utility/api';
 import React, { Component } from 'react';
 import AppHeader from '../components/AppHeader';
 import AdList from '../components/advertisements/AdvertisementList';
-import CreateButton from '../components/CreateButton';
+import CreateButton from '../components/buttons/CreateButton';
 
 export default class UserAdvertisementsPage extends Component {
   constructor(props){
